Update T3 code
Remove PutS as it was not allowing the code to compile. We will have to
figure out why it is not working.

Previous T3 was not working. Update code to make it work by making sure
input is read correctly and ends at EOF. Use i instead of count as i
marks the address where the code would read from.

Bug fixes after testing. 0 was being read for no reason so make sure
it is only read when it is coming from the file.